requirements.txt
appnope==0.1.0
decorator==4.0.6
dj-database-url==0.3.0
Django==1.9.1
djangorestframework==3.3.2
drf-extensions==0.2.8
fake-factory==0.5.3
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
nose==1.3.7
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
